requirements.txt
python-docx
numpy
openai
pandas
plotly
PyPDF2
scikit_learn
streamlit
tensorflow